requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==21.4.0
chardet==4.0.0
charset-normalizer==3.2.0
discord.py==2.3.1
frozenlist==1.4.0
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
typing_extensions==4.2.0
yarl==1.7.2
